ml_pipeline/requirements.txt
torch
torch-geometric
ogb
wandb
tqdm
torch-sparse
torch-scatter
aiohttp
pandas
google-generativeai